src-scraper/init_data/requirements.txt
requests
python-dotenv
unidecode